refactor(api): move validators to evaluation scope
