+++ api/src/evaluation/domain/models/ValidatorQCU.js
@@ -1,5 +1,5 @@
 import * as solutionServiceQCU from '../../../../lib/domain/services/solution-service-qcu.js';
-import { Validation } from './Validation.js';
+import { Validation } from '../../../shared/domain/models/Validation.js';
 import { Validator } from './Validator.js';
 
 /**
