+++ api/src/evaluation/domain/models/ValidatorQROC.js
@@ -1,5 +1,5 @@
 import * as solutionServiceQROC from '../../../../lib/domain/services/solution-service-qroc.js';
-import { Validation } from './Validation.js';
+import { Validation } from '../../../shared/domain/models/Validation.js';
 import { Validator } from './Validator.js';
 
 /**
