+++ api/src/shared/domain/models/index.js
@@ -117,13 +117,6 @@ import { UserDetailsForAdmin } from './UserDetailsForAdmin.js';
 import { UserOrgaSettings } from './UserOrgaSettings.js';
 import { UserSavedTutorialWithTutorial } from './UserSavedTutorialWithTutorial.js';
 import { Validation } from './Validation.js';
-import { Validator } from './Validator.js';
-import { ValidatorAlwaysOK } from './ValidatorAlwaysOK.js';
-import { ValidatorQCM } from './ValidatorQCM.js';
-import { ValidatorQCU } from './ValidatorQCU.js';
-import { ValidatorQROC } from './ValidatorQROC.js';
-import { ValidatorQROCMDep } from './ValidatorQROCMDep.js';
-import { ValidatorQROCMInd } from './ValidatorQROCMInd.js';
 
 export {
   AccountRecoveryDemand,
@@ -249,11 +242,4 @@ export {
   UserSavedTutorialWithTutorial,
   UserToCreate,
   Validation,
-  Validator,
-  ValidatorAlwaysOK,
-  ValidatorQCM,
-  ValidatorQCU,
-  ValidatorQROC,
-  ValidatorQROCMDep,
-  ValidatorQROCMInd,
 };
