+++ api/tests/evaluation/unit/domain/models/ValidatorQROC_test.js
@@ -1,7 +1,7 @@
-import { AnswerStatus } from '../../../../src/shared/domain/models/AnswerStatus.js';
-import { Validation } from '../../../../src/shared/domain/models/Validation.js';
-import { ValidatorQROC } from '../../../../src/shared/domain/models/ValidatorQROC.js';
-import { domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { ValidatorQROC } from '../../../../../src/evaluation/domain/models/ValidatorQROC.js';
+import { AnswerStatus } from '../../../../../src/shared/domain/models/AnswerStatus.js';
+import { Validation } from '../../../../../src/shared/domain/models/Validation.js';
+import { domainBuilder, expect, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Domain | Models | ValidatorQROC', function () {
   let solutionServiceQROCStub;
